refactor: separate fetchFormattedRecord as helper
